binder: Fix ABI incompatibility between older and newer compilers
Parcel::writeString16's first parameter was a char16_t - defined to

uint16_t somewhere else.

With ISO C++11 compatible compilers, however, char16_t is a

compiler-defined-type that is not necessarily unsigned.

This causes the symbol to be different, breaking support for talking
to binary blobs built with pre-ISO C++11-compilers, such as the
Galaxy Nexus RIL.

Change-Id: I3862e3e2b966c618467b198e0197cb3e4024a08a